next-20260907 + simpler KVM patch

+++ fs/ufs/dir.c
@@ -213,7 +213,7 @@ fail:
 static unsigned
 ufs_last_byte(struct inode *inode, unsigned long page_nr)
 {
-	u64 last_byte = inode->i_size;
+	unsigned last_byte = inode->i_size;
 
 	last_byte -= page_nr << PAGE_SHIFT;
 	if (last_byte > PAGE_SIZE)
